+++ packages/moonstone/CHANGELOG.md
@@ -13,6 +13,10 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Input` to not spot its own input decorator when we 5way out.
 - `moonstone/Button` text color when used on a neutral (light) background in some cases
 
+### Removed
+
+- `moonstone/ExpandableInput` property `onInputChange`.
+
 ## [1.0.0] - 2017-03-31
 
 > NOTE: We have also modified most form components to be usable in a controlled (app manages component